[framework] Adjust coin module decimal type (#2935)

+++ aptos-move/framework/aptos-framework/sources/managed_coin.move
@@ -56,7 +56,7 @@ module aptos_framework::managed_coin {
         account: &signer,
         name: vector<u8>,
         symbol: vector<u8>,
-        decimals: u64,
+        decimals: u8,
         monitor_supply: bool,
     ) {
         let (burn_cap, freeze_cap, mint_cap) = coin::initialize<CoinType>(
